Split library
also removed breaker from ircc
